first commit of File-ECG_backward
Purpose: Apply prepared testing set to the system.

Method:
Disable all bluetooth-related functions and simulate the bluetooth ECG
sensor with a ReadFromFile function (Probably to be modified as a object in
the future).

Current Status:
The system can't reconstruct the transmitted testing set properly.
It seems that there's some problems in the QRS Detection and the system
may crash after the first time transmitting a 10*280 jp2 image.